Remove section specific padding for now.
Section specific padding introduces problems across the entire layout
stack for now so Padding layout manager is restricted to recycler view
padding only for now. When full decorator support is added it will be
possible to add selective padding depending on dynamic display
parameters. Which will be better than any section absed padding
configuration anyway. So possibly this problem may never be fixed. In
that case it should be noted that Plms should be made a singleton which
is only used as the initial slm in the entire slm layout stack.

For now, this marks the completion of linear section, inline headers and
footers, sticky footers, and collapsible/expandable sections.

Closes #190, closes #193, closes #195, closes #196, closes #214.